+++ src/app/app.component.ts
@@ -9,7 +9,7 @@ import * as screenfull from 'screenfull';
 
 import { AppSettingsService } from './app-settings.service';
 import { DataSetService } from './data-set.service';
-import { NotificationsService } from './notifications.service';
+import { NotificationsService, activeAlarms } from './notifications.service';
 
 
 declare var NoSleep: any; //3rd party
@@ -24,7 +24,6 @@ export class AppComponent implements OnInit, OnDestroy {
   noSleep = new NoSleep();
 
   pageName: string = '';
-  rootPageIndexSub: Subscription;
 
   unlockStatus: boolean = false;
   unlockStatusSub: Subscription;
@@ -35,7 +34,15 @@ export class AppComponent implements OnInit, OnDestroy {
   themeName: string;
   themeClass: string = 'default-light fullheight';
   themeNameSub: Subscription;
+
   notificationSub: Subscription;
+  alarmSub: Subscription;
+
+  alarms: activeAlarms = {};
+  alarmCount: number = 0;
+  unAckAlarms: number = 0;
+  blinkWarn: boolean = false;
+  blinkCrit: boolean = false;
 
   constructor(
     private AppSettingsService: AppSettingsService,
@@ -63,7 +70,8 @@ export class AppComponent implements OnInit, OnDestroy {
     )
     this.DataSetService.startAllDataSets();
     
-    this.notificationSub = this.NotificationsService.getObservable().subscribe(
+    // Snackbar Notification Code
+    this.notificationSub = this.NotificationsService.getNotificationObservable().subscribe(
       appNotififaction => {
         this._snackBar.open(appNotififaction.message, 'dismiss', {
           duration: appNotififaction.duration,
@@ -71,11 +79,18 @@ export class AppComponent implements OnInit, OnDestroy {
         });
       }
     )
+    // Alarm code
+
+    this.alarmSub = this.NotificationsService.getAlarmObservable().subscribe(
+      alarms  => {
+        this.alarms = alarms;
+        this.updateAlarms();
+       }
+    )
 
   }
 
   ngOnDestroy() {
-    this.rootPageIndexSub.unsubscribe();
     this.unlockStatusSub.unsubscribe();
     this.themeNameSub.unsubscribe();
     this.notificationSub.unsubscribe();
@@ -85,6 +100,7 @@ export class AppComponent implements OnInit, OnDestroy {
   setTheme(theme: string) {
     this.AppSettingsService.setThemName(theme);
   }
+
 
   unlockPage() {
     if (this.unlockStatus) {
@@ -97,13 +113,72 @@ export class AppComponent implements OnInit, OnDestroy {
     this.AppSettingsService.setUnlockStatus(this.unlockStatus);
   }
 
+  updateAlarms() {
+    this.alarmCount = Object.keys(this.alarms).length;
+    this.unAckAlarms = 0;
+
+    if (this.alarmCount > 0) {
+      // find worse alarm state
+      let sev = 0;
+
+      for (const [path, alarm] of Object.entries(this.alarms))
+      {
+        if (alarm.ack) { continue; }
+        let aSev = 0;
+        switch (alarm.state) {
+          case 'alert':
+          case 'warn':
+            aSev = 1;
+            this.unAckAlarms++;
+            break;
+          case 'alarm':
+          case 'emergency':
+            aSev = 2;
+            this.unAckAlarms++;
+        }
+        if (aSev > sev) { sev = aSev; }
+      }
+
+      switch(sev) {
+        case 0:
+          this.blinkWarn = false;
+          this.blinkCrit = false;
+          break;
+        case 1:
+          this.blinkWarn = true;
+          this.blinkCrit = false;
+          break;
+        case 2:
+          this.blinkCrit = true;
+          this.blinkWarn = false;
+
+      }
+    } else {
+      // no Alarms
+      this.blinkWarn = false;
+      this.blinkCrit = false;
+    }
+  }
+
+  ackAlarm(path: string, timeout: number = 0) {
+    if (path in this.alarms) {
+      this.alarms[path].ack = true;
+    }
+    if (timeout > 0) {
+      setTimeout(()=>{
+        console.log(path);
+        if (path in this.alarms) {
+          this.alarms[path].ack = false;
+        }
+        this.updateAlarms();
+      }, timeout);
+    }
+    this.updateAlarms();
+  }
+
   newPage() {
     this.LayoutSplitsService.newRootSplit();
       //this.router.navigate(['/page', rootNodes.findIndex(uuid => uuid == newuuid)]);
-  }
-
-  deletePage() {
-
   }
 
   pageDown() {